Feature/son nh/featute dang ky hoc phan

+++ src/main/java/com/example/university/model/CTDTNhanVien.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class CTDTNhanVien {
     private Integer id;
-    private Integer maCTDT;
+    private Long maCTDT;
     private Integer maNV;
 }
