+++ src/main/java/com/example/university/model/HocPhan.java
@@ -6,11 +6,11 @@ import lombok.Setter;
 @Getter
 @Setter
 public class HocPhan {
-    private Integer maHP;
+    private Long maHP;
     private String tenHP;
     private Integer soTinChi;
     private Integer tietLT;
     private Integer tietTH;
-    private Integer maCTDT;
-    private Integer tienQuyet;
+    private Long maCTDT;
+    private Long tienQuyet;
 }
